fix(github-stats): keep readGitHubStats server-only for client import protection
Moving the footer stats fetch into the (main)/(pure) layout loaders must
not pull cloudflare:workers (via GitHub/shared) into the client bundle —
TanStack's client import protection denies that specifier.

- Move readGitHubStats into stats.server.ts so server.ts references it
  only inside the fetchGitHubStats server-fn handler, which the client
  build strips (mirrors how -stats.server.ts isolates Plausible).
- Route loaders call the fetchGitHubStats server fn (RPC), not the plain
  readGitHubStats, matching the repo's loader convention.
- Rename the test to stats.server.test.ts alongside its module.

Verified: full 5-step vite build now passes the client build step that
previously failed with 'Denied by specifier pattern: cloudflare:workers'.
