+++ src/lib/domains/GitHub/server.ts
@@ -3,63 +3,18 @@ import { createServerFn } from "@tanstack/react-start";
 
 import { DEFAULT_REPO } from "@/config";
 import { GITHUB_API_BASE_URL, getGitHubHeaders, getGitHubToken } from "@/lib/domains/GitHub/shared";
+import { readGitHubStats } from "@/lib/domains/GitHub/stats.server";
 import { type RepoContributor } from "@/lib/domains/GitHub/types";
 
-export type GitHubStats = { stars: number; forks: number };
+export { type GitHubStats } from "@/lib/domains/GitHub/stats.server";
 
-function isGitHubStats(value: unknown): value is GitHubStats {
-	return (
-		typeof value === "object" &&
-		value !== null &&
-		"stars" in value &&
-		"forks" in value &&
-		typeof value.stars === "number" &&
-		typeof value.forks === "number"
-	);
-}
-
-const STATS_CACHE_NAME = "github-stats";
-const STATS_CACHE_KEY = "https://internal.cache/github-stats";
-const STATS_CACHE_TTL = 3600;
-
-export async function readGitHubStats(): Promise<GitHubStats> {
-	const cache = await caches.open(STATS_CACHE_NAME);
-	const cacheKey = new Request(STATS_CACHE_KEY);
-	const cached = await cache.match(cacheKey);
-	if (cached) {
-		const data: unknown = await cached.json();
-		if (isGitHubStats(data)) return data;
-	}
-
-	const token = getGitHubToken();
-	const request = await fetch(
-		`${GITHUB_API_BASE_URL}/repos/${DEFAULT_REPO.owner}/${DEFAULT_REPO.repo}`,
-		{
-			headers: getGitHubHeaders(token),
-		},
-	);
-
-	if (!request.ok) {
-		return { stars: 0, forks: 0 };
-	}
-
-	const data: Endpoints["GET /repos/{owner}/{repo}"]["response"]["data"] = await request.json();
-
-	const { stargazers_count: stars, forks_count: forks } = data;
-	const stats: GitHubStats = { stars, forks };
-
-	await cache.put(
-		cacheKey,
-		new Response(JSON.stringify(stats), {
-			headers: {
-				"content-type": "application/json",
-				"cache-control": `max-age=${STATS_CACHE_TTL}`,
-			},
-		}),
-	);
-
-	return stats;
-}
+/**
+ * Server-fn wrapper so route loaders call the stats read over RPC. Route modules
+ * import this (not `readGitHubStats`) so the client build strips the handler and
+ * its server-only imports — `readGitHubStats` lives in `stats.server.ts`, whose
+ * `cloudflare:workers` dependency must never reach the client bundle.
+ */
+export const fetchGitHubStats = createServerFn({ method: "GET" }).handler(() => readGitHubStats());
 
 export const fetchRepoContributors = createServerFn({ method: "GET" }).handler(async () => {
 	const token = getGitHubToken();
