+++ src/routes/(pure)/route.tsx
@@ -3,11 +3,11 @@ import { createFileRoute, Outlet } from "@tanstack/react-router";
 import { Footer } from "@/lib/components/Footer";
 import { Header } from "@/lib/components/Header";
 import { FoobarPixel } from "@/lib/domains/foobar/Pixel";
-import { readGitHubStats } from "@/lib/domains/GitHub/server";
+import { fetchGitHubStats } from "@/lib/domains/GitHub/server";
 
 export const Route = createFileRoute("/(pure)")({
 	component: MainLayout,
-	loader: () => readGitHubStats().catch(() => ({ stars: 0, forks: 0 })),
+	loader: () => fetchGitHubStats().catch(() => ({ stars: 0, forks: 0 })),
 });
 
 function MainLayout() {
